+++ test/e2e/kind_sweep_webgl.jl
@@ -180,7 +180,8 @@ HTML(
 # ╔═╡ d1000000-0000-0000-0000-000000000040
 HTML(
     "<span id=\"kind_meta\" style=\"display:none\">$(JSON3.write(kind_sweep_meta()))</span>" *
-        "<span id=\"kind_backend\">webgl</span>",
+        "<span id=\"kind_backend\">webgl</span>" *
+        "<span id=\"kind_env\" style=\"display:none\">$(!isempty(dev))</span>",
 )
 
 # ╔═╡ Cell order: